2026年3月1日 01:47

+++ CSP-J yusiwen/ngs/ngs.cpp
@@ -3,7 +3,7 @@
 using namespace std;
 #define el '\n'
 const ll N=1e5+5;
-ll a,b,c;
+double a,b,c;
 void solve(){
     cin>>a>>b>>c;
     cout<<(c-a)*((c-b)/(b-a))+c;
